fix!: Allow cluster deletion by default
Adds an option flag to control `deletion_prevention` with `false` as the
default value. This was introduced in Google provider v5 to prevent
accidental deletion of clusters. There are no functional changes to the
`sa` or `kubeconfig` modules but they also include the same provider
constraints for consistency between all modules in repo.

BREAKING-CHANGE: The module requires Google provider v5.21+.
Closes #52.